requirements.txt
streamlit==1.31.0
opencv-python==4.8.1.78
mediapipe==0.10.9
numpy==1.26.3
google-generativeai==0.3.2
openpyxl==3.1.2
pyttsx3==2.90
Pillow==10.2.0
